pyproject.toml
[project]
name = "yawiki"
version = "0.1.0"
description = "Yandex Wiki MCP Server and CLI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=3.0.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[project.scripts]
yawiki = "yawiki.mcp_server:mcp.run"
yawiki-cli = "yawiki.client:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/yawiki"]
